:recycle: (general) Move various types into their own assemblies
Doesn't build yet
