+++ src/MetaCraft.Archive/AssemblyExportDeclaration.cs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.Text.Json.Serialization;
-using MetaCraft.Core.Serialization;
+using MetaCraft.Common.Json;
 using Semver;
 
 namespace MetaCraft.Core.Manifest;
