+++ src/MetaCraft.Archive/ManifestRuntimeInfo.cs
