+++ src/MetaCraft.Common/Json/SemVersionConverter.cs
@@ -3,10 +3,9 @@
 
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using MetaCraft.Core.Locales;
 using Semver;
 
-namespace MetaCraft.Core.Serialization;
+namespace MetaCraft.Common.Json;
 
 public sealed class SemVersionConverter : JsonConverter<SemVersion>
 {
@@ -20,7 +19,8 @@ public sealed class SemVersionConverter : JsonConverter<SemVersion>
 
         if (!SemVersion.TryParse(strValue, out var result))
         {
-            throw new JsonException(Lc.L("The provided string is not a valid Semantic Version number."));
+            throw new JsonException("The provided string is not a valid Semantic Version number.");
+            // TODO: add a common Lc module and restore this localization string
         }
         return result;
     }
