+++ src/MetaCraft.Common/Platform/PlatformIdentifier.cs
@@ -3,12 +3,12 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.JavaScript;
 using System.Text;
 using System.Text.Json.Serialization;
-using MetaCraft.Core.Locales;
-using MetaCraft.Core.Serialization;
+using MetaCraft.Common.Json;
 
-namespace MetaCraft.Core.Platform;
+namespace MetaCraft.Common.Platform;
 
 [JsonConverter(typeof(PlatformIdentifierConverter))]
 public readonly record struct PlatformIdentifier
@@ -48,10 +48,10 @@ public readonly record struct PlatformIdentifier
     /// <para>
     /// This platform "includes" the other platform identifier if any of the following conditions are true:
     /// <list type="bullet">
-    ///     <item>This and the other platform identifier equals (that is, <see cref="Equals(MetaCraft.Core.Platform.PlatformIdentifier?)"/> returns true).</item>
-    ///     <item>Both <see cref="System"/> and <see cref="Architecture"/> properties of this instance are correspondingly <see cref="PlatformSystem.Any"/> and <see cref="PlatformArchitecture.Any"/>.</item>
-    ///     <item><see cref="System"/> is <see cref="PlatformSystem.Any"/>, and <see cref="Architecture"/> property on both instances are the same.</item>
-    ///     <item><see cref="Architecture"/> is <see cref="PlatformArchitecture.Any"/>, and the <see cref="System"/> property on both instances are the same.</item>
+    ///     <item>This and the other platform identifier equals (that is, <see cref="Equals(PlatformIdentifier?)"/> returns true).</item>
+    ///     <item>Both <see cref="System"/> and <see cref="Architecture"/> properties of this instance are correspondingly <see cref="JSType.Any"/> and <see cref="JSType.Any"/>.</item>
+    ///     <item><see cref="System"/> is <see cref="JSType.Any"/>, and <see cref="Architecture"/> property on both instances are the same.</item>
+    ///     <item><see cref="Architecture"/> is <see cref="JSType.Any"/>, and the <see cref="System"/> property on both instances are the same.</item>
     /// </list>
     /// </para>
     /// </remarks>
@@ -149,10 +149,10 @@ public readonly record struct PlatformIdentifier
     {
         var message = result switch
         {
-            ParseFunctionResult.InvalidFormat => Lc.L("Invalid platform identifier format."),
-            ParseFunctionResult.InvalidSystem => Lc.L("Invalid system identifier."),
-            ParseFunctionResult.InvalidArchitecture => Lc.L("Invalid architecture identifier."),
-            _ => Lc.L("Unexpected error occurred when parsing the platform identifier: '{0}'", result)
+            ParseFunctionResult.InvalidFormat => "Invalid platform identifier format.",
+            ParseFunctionResult.InvalidSystem => "Invalid system identifier.",
+            ParseFunctionResult.InvalidArchitecture => "Invalid architecture identifier.",
+            _ => string.Format("Unexpected error occurred when parsing the platform identifier: '{0}'", result)
         };
 
         return new FormatException(message);
