+++ src/MetaCraft.Common/SemVersionKey.cs
@@ -2,9 +2,10 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.Text.Json.Serialization;
+using MetaCraft.Common.Json;
 using Semver;
 
-namespace MetaCraft.Core.Serialization;
+namespace MetaCraft.Common;
 
 /// <summary>
 /// Represents <see cref="SemVersion"/> as a key in a dictionary.
