+++ src/MetaCraft.Core/CoreJsonContext.cs
@@ -2,11 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.Text.Json.Serialization;
+using MetaCraft.Archive.References;
+using MetaCraft.Common;
+using MetaCraft.Common.Platform;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Manifest;
 using MetaCraft.Core.Platform;
 using MetaCraft.Core.Scopes.Referral;
-using MetaCraft.Core.Serialization;
 
 namespace MetaCraft.Core;
