+++ src/MetaCraft.Core/Dependency/DependencyChecker.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2024 WithLithum <WithLithum@outlook.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+using MetaCraft.Archive.References;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Locales;
 using MetaCraft.Core.Scopes;
