+++ src/MetaCraft.Core/Transactions/PackageInstallTransaction.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2024 WithLithum <WithLithum@outlook.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+using MetaCraft.Common.Platform;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Dependency;
 using MetaCraft.Core.Locales;
