+++ src/MetaCraft.Core/Transactions/UpdateReferrersTransaction.cs
@@ -1,11 +1,12 @@
 // SPDX-FileCopyrightText: 2024 WithLithum <WithLithum@outlook.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+using MetaCraft.Archive.References;
+using MetaCraft.Common;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Locales;
 using MetaCraft.Core.Scopes;
 using MetaCraft.Core.Scopes.Referral;
-using MetaCraft.Core.Serialization;
 using Semver;
 
 namespace MetaCraft.Core.Transactions;
