+++ src/MetaCraft.Tests/DependencyTests.cs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: 2024 WithLithum <WithLithum@outlook.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+using MetaCraft.Archive.References;
 using MetaCraft.Core;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Dependency;
