+++ src/MetaCraft.Tests/SemVersionKeyTests.cs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 WithLithum <WithLithum@outlook.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-using MetaCraft.Core.Serialization;
+using MetaCraft.Common;
 using Semver;
 
 namespace MetaCraft.Tests;
