+++ src/MetaCraft/Commands/InstallCommand.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 using System.CommandLine;
+using MetaCraft.Common.Platform;
 using MetaCraft.Core;
 using MetaCraft.Core.Archive;
 using MetaCraft.Core.Dependency;
